requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
dataclasses-json==0.6.7
dnspython==2.8.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.5
idna==3.10
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-mongodb==0.7.0
langchain-text-splitters==0.3.11
langchain-voyageai==0.1.7
langsmith==0.4.27
lark==1.2.2
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.15.1
pypdf==6.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.7.2
scipy==1.16.2
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
voyageai==0.3.5
yarl==1.20.1
zstandard==0.25.0
